+++ UltraPortal/Constants.cs
@@ -9,10 +9,15 @@ namespace UltraPortal {
         public static LayerMask PortalLayer => LayerMask.NameToLayer("Portal");
 
         public const string PortalExplosionWeapon = "dynamicportal";
+        public const string PortalProjectileWeapon = "portalprojectile";
 
         public const string StyleSafetyHazardId = "style.ultraportal.safetyhazard";
         public const string StyleSafetyHazardName = "SAFETY HAZARD";
-        public const int StyleSafetyHazardPoints = 500;
+        public const int StyleSafetyHazardPoints = 100;
+        
+        public const string StylePortalProjectileId = "style.ultraportal.projectiled";
+        public const string StylePortalProjectileName = "DISPLACEMENT";
+        public const int StylePortalProjectilePoints = 25;
         
 
         private static Camera _mainCamera;
@@ -51,12 +56,14 @@ namespace UltraPortal {
             
             public const string  PortalBundle = "portals";
             public const string  WeaponBundle = "weapons";
+            public const string  DebugBundle = "debug";
 
-            public const string  PortalExit = "Portal Exit";
-            public const string  Mirror = "Mirror";
+            public const string PortalExit = "Portal Exit";
+            public const string Mirror = "Mirror";
 
             public const string PortalGun = "Portal Gun";
             public const string MirrorGun = "Mirror Gun";
+            public const string TwistGun = "Twist Gun Variant";
             
             public const string Projectile = "Projectile";
 
@@ -64,6 +71,8 @@ namespace UltraPortal {
             public const string PortalGunIconGlow = "UltraPortalGunIconGlow";
 
             public const string Explosion = "Explosion";
+
+            public const string DebugForwardArrow = "ForwardArrow";
         }
     }
 }