+++ UltraPortal/DynamicPortalExit.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using Interop.std;
 using ULTRAKILL.Portal;
 using UltraPortal.Colorizers;
 using UltraPortal.Extensions;
@@ -20,9 +21,10 @@ namespace UltraPortal {
 
 		private const float SphereCheckRadius = 0.2f;
 
-		private static Action<PortalSide, Collider, bool, bool> _toggleColliderAction;
+		private Action<PortalSide, Collider, bool, bool> _toggleColliderAction;
 
 		public Action OnInitialized;
+		public DynamicPortalExit otherExit;
 		
 		public bool IsEntityNear {
 			get {
@@ -41,17 +43,16 @@ namespace UltraPortal {
 				_playerNearExit = value;
 			}
 		}
-		
+
+		public bool IsInitialized => (transform.position - PortalGunBase.DefaultPortalPosition).sqrMagnitude > 1;
+
 		// PORTAL
 		public Portal hostPortal;
 		public PortalSide side;
 		public PortalGunBase hostGun;
 		
 		public bool AssistedPortalTravel { get; private set; } = false;
-
-		// COLLIDERS
-		private DynamicPortalTooClose _tooClose;
-
+		
 		private GameObject _passableBlockage;
 		
 		public Vector3 PortalCenter {
@@ -97,20 +98,7 @@ namespace UltraPortal {
 			GameObject keepActive = new GameObject($"{name} Keep Active");
 			_keepActive = keepActive.AddComponent<KeepActive>();
 			
-			// Get too close trigger
-			Transform tooCloseTransform = transform.Find(TooCloseTrigger);
-			if (!tooCloseTransform) {
-				Plugin.LogSource.LogError($"Failed to find {TooCloseTrigger} on {name}.");
-				return;
-			}
-
-			_tooClose = tooCloseTransform.gameObject.AddComponent<DynamicPortalTooClose>();
-			
 			_toggleColliderAction += (portalSide, collider, toggle, assistance) => {
-				// if (assistedPortalTravel && portalSide != side) {
-				// 	return;
-				// }
-				
 				ToggleColliders(toggle, collider, assistance);
 			};
 		}
@@ -120,6 +108,21 @@ namespace UltraPortal {
 				return;
 			}
 			
+			if (c.transform.IsChildOf(transform)) {
+				return;
+			}
+			
+			if (!AssistedPortalTravel) {
+				Vector3 dir = (c.transform.position - transform.position).normalized;
+				float dot = Vector3.Dot(-transform.forward, dir);
+				
+				LogInfo($"Dot of {c.name}: {dot}");
+				if (Mathf.Abs(dot) > ModConfig.PerpendicularThreshold.GetValue()) {
+					LogVerboseInfo($"Rejected: {c.name} (close to perpendicular to exit)");
+					return;
+				}
+			}
+
 			LogVerboseInfo($"Adding collider: {c.name}; Checking children: {checkChildren}");
 			
 			_colliders.SafeAdd(c);
@@ -145,7 +148,7 @@ namespace UltraPortal {
 			
 			// Raycast to ensure some slopes (like at the start of 8-2) work
 			RaycastHit[] sphereCastResults = Physics.SphereCastAll(transform.position, SphereCheckRadius,
-				transform.forward, 3f, EnvironmentLayer, QueryTriggerInteraction.Ignore);
+				-transform.forward, 3f, EnvironmentLayer, QueryTriggerInteraction.Ignore);
 			if (sphereCastResults.Length < 1) {
 				return;
 			}
@@ -161,8 +164,9 @@ namespace UltraPortal {
 
 		private void CalculateAssistance() {
 			// Check if portal is facing upwards
-			float dot = Mathf.Abs(Vector3.Dot(transform.forward, NewMovement.Instance.rb.GetGravityVector()));
-			AssistedPortalTravel = dot > 0.6f;
+			float dot = Mathf.Abs(Vector3.Dot(transform.forward.normalized, NewMovement.Instance.rb.GetGravityVector().normalized));
+			LogVerboseInfo($"{name} dot to {NewMovement.Instance.rb.GetGravityVector().normalized}: {dot}");
+			AssistedPortalTravel = dot > ModConfig.AssistedPortalThreshold.GetValue();
 		}
 
 		private void Cleanup() {
@@ -215,7 +219,9 @@ namespace UltraPortal {
 			
 			_colorManager.ColorPortal();
 			_keepActive.target = gameObject;
-
+      
+      ShowForwardArrow(transform.position, -transform.forward, 10f);
+      
 			if (!AudioManager.Instance) {
 				LogError("Audio Manager is not present in scene!");
 				return;
@@ -291,6 +297,10 @@ namespace UltraPortal {
 			
 			CalculateAssistance();
 			
+			if (otherExit) {
+				otherExit.CalculateAssistance();
+				otherExit._toggleColliderAction.Invoke(side, other, true, false);
+			}
 			_toggleColliderAction.Invoke(side, other, true, AssistedPortalTravel);
 		}
 
@@ -307,18 +317,38 @@ namespace UltraPortal {
 			if (!hostGun || !hostPortal) {
 				return;
 			}
-			
-			if (hostGun is PortalGun portalGun) {
-				switch (side) {
-					case PortalSide.Enter:
+
+			switch (hostGun.variant) {
+				case WeaponVariant.BlueVariant:
+					PortalGun portalGun = hostGun as PortalGun;
+					
+					if (side == PortalSide.Enter) {
 						portalGun.SpawnEntry(true);
-						break;
-					case PortalSide.Exit:
+					}
+					else {
 						portalGun.SpawnExit(true);
-						break;
-				}
-			} else if (hostGun is MirrorGun mirrorGun) {
-				mirrorGun.SpawnPrimaryMirror(true);
+					}
+					break;
+				case WeaponVariant.GreenVariant:
+					MirrorGun mirrorGun = hostGun as MirrorGun;
+					
+					if (side == PortalSide.Enter) {
+						mirrorGun.SpawnPrimaryMirror(true);
+					}
+					else {
+						mirrorGun.SpawnFlippedMirror(true);
+					}
+					break;
+				case WeaponVariant.RedVariant:
+					TwistGun twistGun = hostGun as TwistGun;
+					
+					if (side == PortalSide.Enter) {
+						twistGun.SpawnEntry(true);
+					}
+					else {
+						twistGun.SpawnExit(true);
+					}
+					break;
 			}
 		}
 
@@ -357,7 +387,11 @@ namespace UltraPortal {
 				if(_passableBlockage.activeSelf)
 					return;
 			}
-			
+
+			if (otherExit) {
+				otherExit.CalculateAssistance();
+				otherExit._toggleColliderAction.Invoke(side, other, false, otherExit.AssistedPortalTravel);
+			}
 			_toggleColliderAction.Invoke(side, other, false, AssistedPortalTravel);
 			_currentTravellers.SafeRemove(other);
 		}
@@ -377,12 +411,14 @@ namespace UltraPortal {
 				
 				NewMovement.Instance.GetComponent<KeepInBounds>().enabled = !value;
 				NewMovement.Instance.GetComponent<WallCheckGroup>().enabled = !value;
-				return;
 			}
 
 			EnemyIdentifier eid = other.GetComponent<EnemyIdentifier>();
 			if (eid) {
-				other.attachedRigidbody.detectCollisions = !value;
+				if (!assisted) {
+					value = true;
+				}
+				other.attachedRigidbody.detectCollisions = value;
 			}
 		}