+++ UltraPortal/Patches/EnemyPatches.cs
@@ -2,29 +2,39 @@ using System.Collections.Generic;
 using HarmonyLib;
 using UnityEngine;
 using static UltraPortal.Constants;
+using static UltraPortal.DebugUtils;
 
 namespace UltraPortal {
 	[HarmonyPatch]
 	public static class EnemyPatches {
-		public static List<EnemyIdentifier> AlreadyDealtWith = new List<EnemyIdentifier>();
+		public static List<EnemyIdentifier> AlreadyAppliedStyle = new List<EnemyIdentifier>();
+
+		private static void ApplyStyleBonus(EnemyIdentifier eid, string id, int points, Color color) {
+			LogVerboseInfo($"using style: {id}");
+			StyleHUD.Instance.AddPoints(points, id,
+				prefix: $"<color=#{ColorUtility.ToHtmlStringRGB(color)}>",
+				postfix: "</color>");	
+			
+			AlreadyAppliedStyle.Add(eid);
+		}
 		
 		[HarmonyPostfix]
 		[HarmonyPatch(typeof(EnemyIdentifier), nameof(EnemyIdentifier.DeliverDamage))]
 		public static void DeliverDamagePatch(EnemyIdentifier __instance) {
-			if (AlreadyDealtWith.Contains(__instance)) {
+			if (AlreadyAppliedStyle.Contains(__instance)) {
 				return;
 			}
 			
-			Plugin.LogSource.LogInfo($"checking if enemy qualifies for safety hazard; name: {__instance.name}");
+			LogVerboseInfo($"checking if enemy qualifies for style bonuses; name: {__instance.name}");
 			
 			if (__instance.hitterWeapons.Contains(PortalExplosionWeapon)) {
-				Plugin.LogSource.LogInfo("using safety hazard style!");
-				StyleHUD.Instance.AddPoints(StyleSafetyHazardPoints, StyleSafetyHazardId,
-					prefix: $"<color=#{ColorUtility.ToHtmlStringRGB(ModConfig.SafetyHazardColor.GetValue())}>",
-					postfix: "</color>");
+				ApplyStyleBonus(__instance, StyleSafetyHazardId, StyleSafetyHazardPoints, ModConfig.SafetyHazardColor.GetValue());
 			}
-			
-			AlreadyDealtWith.Add(__instance);
+
+			if (__instance.hitterWeapons.Contains(PortalProjectileWeapon)) {
+				ApplyStyleBonus(__instance, StylePortalProjectileId, StylePortalProjectilePoints,
+					ModConfig.ProjectileBonusColor.GetValue());
+			}
 		}
 	}
 }