+++ UltraPortal/Plugin.cs
@@ -8,6 +8,7 @@ using BepInEx.Logging;
 using Configgy;
 using HarmonyLib;
 using ULTRAKILL.Portal.Geometry;
+using UltraPortal.Projectiles;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
@@ -21,7 +22,7 @@ namespace UltraPortal {
         private static class PluginInfo {
             public const string Name = "ULTRAPORTAL";
             public const string Guid = "com.ultraportal";
-            public const string Version = "0.1.1";
+            public const string Version = "0.2.0";
         }
         
         public static ManualLogSource LogSource { get; private set; }
@@ -49,7 +50,8 @@ namespace UltraPortal {
         }
 
         private void OnSceneLoaded(Scene scene, LoadSceneMode loadMode) {
-            PortalGunManager.EquippedPortalGun = false;
+            PortalGunManager.UsedPortalGun = false;
+            PortalProjectileHelper.PortalScaleSceneStart = ModConfig.PortalScaleMod.GetValue();
             
             if (loadMode != LoadSceneMode.Single || SceneHelper.CurrentScene == "Intro" || SceneHelper.CurrentScene == "Main Menu") {
                 return;
@@ -66,6 +68,7 @@ namespace UltraPortal {
                 
                 // Register styles
                 StyleHUD.Instance.RegisterStyleItem(StyleSafetyHazardId, StyleSafetyHazardName);
+                StyleHUD.Instance.RegisterStyleItem(StylePortalProjectileId, StylePortalProjectileName);
             }
             catch {
                 Logger.LogError("Scene is not compatible! Failed to spawn portal spawner!");