+++ Thunderstore/CHANGELOG.md
@@ -5,9 +5,27 @@
 - Ground portals now detect the proper colliders in the cybergrind. ([see issue](https://github.com/averyocean65/UltraPortal/issues/14))
 - Portals now reset colliders when they're destroyed.
 - Increased size of portal gun projectile to decrease collision errors at higher projectile speeds.
+- *+SAFETY HAZARD* now gets shown reliably.
+- Fixed portal duplication with frozen sandbox objects.
+- Fixed leaderboard submission not being handled properly.
 
-## Config
-- Renamed `Advanced` sub-categories to `Experimental` to indicate that tweaking the values may cause significant bugs.
+### Gameplay
+- Added the **Twist Gun**.
+- Changed *+SAFETY HAZARD* from 500 points to 100 points.
+- Added new style bonus *+DISPLACEMENT*.
+- Added alternative fire for mirror gun.
+- *+USED PORTAL GUN* in the level end screen is only enabled when a portal gun was actually fired.
+- Buffed damage of portal projectile.
+- Portal projectiles only teleport light enemies now.
+
+### Config
+- Changed categorization of config values
+  - Renamed `Advanced` sub-categories to `Experimental` to indicate that tweaking the values may cause significant bugs.
+  - Moved portal color properties to `Visuals/Portals`.
+- Added individual colors for all portals.
+- Added setting to disable portal borders.
+- Added setting to change max recursions of portals.
+- Added setting to scale portals.
 - Added `Debug` category to configuration.
 - Added ability to hide portal borders.
 - Added `Audio` category to configuration.
@@ -18,7 +36,9 @@
   - Portal Close
 - Added ambiance sound effect for (custom) portals.
 
+### Visuals
+- Improved portal and portal gun designs by [RoseWarbug](https://bsky.app/profile/rosewarbug.bsky.social).
+  - Voronoi Shader for portal blockage taken from [ronja-tutorials](https://github.com/ronja-tutorials/ShaderTutorials/tree/master/Assets/028_Voronoi_Noise).
+
 ## Known Bugs
 - It is possible to trigger a portal from behind, which causes things like wall andi floor collisions to disable prematurely.
-- Enemies travelling through portals is unreliable.
-  - Makes it impossible to achieve style modifiers such as *+TERMINAL VELOCITY*.